Move compute_last_updated in addons/utils.py
